Bugs in the sorting section have been fixed. Total Records have been changed to accept long values.

+++ RealEstateApiRepositories/Contacts/IDataResult.cs
@@ -8,6 +8,6 @@ namespace RealEstateApiRepositories.Contacts
     public interface IDataResult<out T> : IResult
     {
         T Data { get; }
-        int TotalRecords { get; }
+        long TotalRecords { get; }
     }
 }
