+++ RealEstateApiRepositories/PaginationDataResult.cs
@@ -25,7 +25,7 @@ namespace RealEstateApiRepositories
         public Uri? FirstPage { get; set; }
         public Uri? LastPage { get; set; }
         public int TotalPages { get; set; }
-        public int TotalRecords { get; set; }
+        public long TotalRecords { get; set; }
         public Uri? NextPage { get; set; }
         public Uri? PreviousPage { get; set; }
         public IReadOnlyList<T> Data { get; set; }
